% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation.R, R/function3.R
\name{correlation}
\alias{correlation}
\title{A NBA player filter function}
\usage{
correlation(year)

correlation(year)
}
\arguments{
\item{year}{indicates the year where we want to convert all numeric variables into a correlation matrix}
}
\description{
This function allows you to return a correlation matrix for all numeric variables in the data within that year

This function allows you to return a correlation matrix for all numeric variables in the data within that year
}
\examples{
correlation()
correlation()
}
\keyword{correlation}
\keyword{matrix}
