% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function2.R, R/oldest_player.R
\name{oldest_player}
\alias{oldest_player}
\title{A NBA player filter function}
\usage{
oldest_player(year)

oldest_player(year)
}
\arguments{
\item{year}{indicates the year where we want to extract the oldest player's information}
}
\description{
This function allows you to return the oldest player's information within a particular season

This function allows you to return the oldest player's information within a particular season
}
\examples{
oldest_player()
oldest_player()
}
\keyword{oldest}
\keyword{player}
