% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function1.R, R/player_pts.R
\name{player_pts}
\alias{player_pts}
\title{A NBA player filter function}
\usage{
player_pts(year)

player_pts(year)
}
\arguments{
\item{year}{indicates the year from the season where a player scores the highest during that season}
}
\description{
This function allows you to return the maximum points scored by each player within a season year

This function allows you to return the maximum points scored by each player within a season year
}
\examples{
player_pts()
player_pts()
}
\keyword{by}
\keyword{each}
\keyword{player}
\keyword{points}
\keyword{scored}
